[IMP] project: replace and use methods overriden in view controllers
Prior to this commit:

Delete conformation dialog props in list and form
views were not easily extendable so in order to change any thing
in that props we have to re write whole dialog props with our
required props

Post this commit:

This commit makes dialog easily extendable so customization of
view can easily extend this props like we do for other
delete confirmation props

Task-3660948

closes odoo/odoo#148046

+++ addons/project/static/src/views/project_task_list/project_task_list_controller.js
@@ -1,19 +1,24 @@
 /** @odoo-module */
 
 import { ListController } from "@web/views/list/list_controller";
-import { DeleteSubtasksConfirmationDialog } from "@project/components/delete_subtasks_confirmation_dialog/delete_subtasks_confirmation_dialog";
+import { subTaskDeleteConfirmationMessage } from "@project/views/project_task_form/project_task_form_controller";
 
 export class ProjectTaskListController extends ListController {
-    async onDeleteSelectedRecords() {
-        if (!Math.max(...this.model.root.selection.map((record) => record.data.subtask_count ))) {
-            return super.onDeleteSelectedRecords();
+
+    get deleteConfirmationDialogProps() {
+        const deleteConfirmationDialogProps = super.deleteConfirmationDialogProps;
+        const hasSubtasks = this.model.root.selection.some(task => task.data.subtask_count > 0)
+        if (!hasSubtasks) {
+            return deleteConfirmationDialogProps;
         }
-        this.dialogService.add(DeleteSubtasksConfirmationDialog, {
+        return {
+            ...deleteConfirmationDialogProps,
             confirm: async () => {
                 await this.model.root.deleteRecords();
                 // A re-load is needed to remove deleted sub-tasks from the view
                 await this.model.load();
             },
-        });
+            body: subTaskDeleteConfirmationMessage,
+        }
     }
 }
